renamed queue_invite entity
