+++ packages/server/ormconfig.ts
@@ -30,7 +30,7 @@ import { UserTokenModel } from './src/profile/user-token.entity';
 import { ChatTokenModel } from './src/chatbot/chat-token.entity';
 import { StudentTaskProgressModel } from './src/studentTaskProgress/studentTaskProgress.entity';
 import { ApplicationConfigModel } from './src/config/application_config.entity';
-import { QueueInviteModel } from './src/queue/queue_invite.entity';
+import { QueueInviteModel } from './src/queue/queue-invite.entity';
 
 config();
