+++ packages/server/src/queue/queue.service.ts
@@ -23,7 +23,7 @@ import { QueueModel } from './queue.entity';
 import { AlertsService } from '../alerts/alerts.service';
 import { ApplicationConfigService } from 'config/application_config.service';
 import { QuestionTypeModel } from 'questionType/question-type.entity';
-import { QueueInviteModel } from './queue_invite.entity';
+import { QueueInviteModel } from './queue-invite.entity';
 
 /**
  * Get data in service of the queue controller and SSE
